[ testsuite2 ] find TYPE/INTERPRETER ERROR anywhere in stdout/stderr
The previous stricter check did not allow to use Java's
RuntimeException for signaling interpreter errors.

The new more lenient check does this, as long as INTERPRETER ERROR is
part of the exception message.

Also, if the interpreter crashes without printing correctly
INTERPRETER ERROR, we shouldn't say that it "ran to completion without
error", rather, we should point to the missing output of INTERPRETER ERROR.